requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.1
httpx>=0.27.0
feedparser>=6.0.10
gunicorn>=21.2.0
psutil>=5.9.8
tqdm>=4.66.0

# Agents stack
openai-agents
openai
pydantic>=2.11,<3
celery>=5.3
redis>=5.0

